refine-fastapi-builder/pyproject.toml
[project]
name = "refine-fastapi-builder"
version = "0.1.0"
description = "Student-Course Management API with layered architecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pwdlib[argon2]>=0.2.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
]
